% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\docType{data}
\name{adult_rxc}
\alias{adult_rxc}
\title{Adult Rx Condition Factors - Table 9}
\format{\describe{
\item{used}{Whether or not the category is used in the RA model}
\item{plat, gold, silver, bronze, cat}{HCC grouping risk score for each
metal level}
\item{rxc}{RXC label for risk scores (Includes leading zeroes)}
}}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
adult_rxc
}
\description{
A dataset containing the adult model prescription drug category (RXC)
factors for the risk adjustment model. Rx condition categories are
triggered by NDC codes (See \code{\link{ndc_rxc}}) or HCPCS codes (See
\code{\link{hcpcs_rxc}}), and a hierarchy is then applied (See
\code{\link{rxc_hier}}).
}
\seealso{
Other factor datasets: 
\code{\link{adult_demo}},
\code{\link{adult_enroll_dur}},
\code{\link{adult_group}},
\code{\link{adult_hcc}},
\code{\link{adult_interaction}},
\code{\link{adult_rxc_hcc_inter}},
\code{\link{child_demo}},
\code{\link{child_group}},
\code{\link{child_hcc}},
\code{\link{infant_demo}},
\code{\link{infant_mat_sev}}
}
\concept{factor datasets}
\keyword{datasets}
