% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_hier}
\alias{cc_hier}
\title{Condition Category Hierarchies - Table 4}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 149 rows and 3 columns.}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
cc_hier
}
\description{
Condition Category Hierarchies - Table 4
}
\keyword{datasets}
