% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_int.R
\docType{data}
\name{cc_int_h}
\alias{cc_int_h}
\title{HCC to interaction group "H" mapping - Table 6}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with
9 rows and 3 columns
\describe{
\item{cc}{Hierarchical condition category (Currently includes some G*)}
\item{var}{Variable mapped to (int_group_h)}
\item{desc}{Short description of the variable}
}
}
\source{
Data import and cleaning at:
\url{https://github.com/EeethB/edgedata/tree/main/data-raw}
}
\usage{
cc_int_h
}
\description{
A dataset containing the mapping from HCC to interaction group "H".
}
\details{
Interaction group H (high?) takes priority over interaction group M
(moderate?). If a member qualifies for group H, they should not get credit
for group M.
}
\seealso{
Other Severe interaction tables: 
\code{\link{cc_int_m}},
\code{\link{cc_severe}}
}
\concept{Severe interaction tables}
\keyword{datasets}
