% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_int.R
\docType{data}
\name{cc_int_m}
\alias{cc_int_m}
\title{HCC to interaction group "M" mapping - Table 6}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with
7 rows and 3 columns
\describe{
\item{cc}{Hierarchical condition category (Currently includes some G*)}
\item{var}{Variable mapped to (int_group_m)}
\item{desc}{Short description of the variable}
}
}
\source{
Most recent DIY entry at:
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index} \cr
Data import and cleaning at:
\url{https://github.com/EeethB/edgedata/tree/master/data-raw}
}
\usage{
cc_int_m
}
\description{
A dataset containing the mapping from HCC to interaction group "M".
}
\details{
Interaction group H (high?) takes priority over interaction group M
(moderate?). If a member qualifies for group H, they should not get credit
for group M.
}
\seealso{
Other Severe interaction tables: 
\code{\link{cc_int_h}},
\code{\link{cc_severe}}
}
\concept{Severe interaction tables}
\keyword{datasets}
