% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_int.R
\docType{data}
\name{cc_severe}
\alias{cc_severe}
\title{HCC to severity group mapping - Table 6}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with
8 rows and 3 columns
\describe{
\item{cc}{Hierarchical condition category (Currently includes some G*)}
\item{var}{Variable mapped to (severe_v3)}
\item{desc}{Short description of the variable}
}
}
\source{
Most recent DIY entry at:
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance} \cr
Data import and cleaning at:
\url{https://github.com/EeethB/edgedata/tree/master/data-raw}
}
\usage{
cc_severe
}
\description{
A dataset containing the mapping from HCC to severe status.
}
\seealso{
Other Severe interaction tables: 
\code{\link{cc_int_h}},
\code{\link{cc_int_m}}
}
\concept{Severe interaction tables}
\keyword{datasets}
