% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\docType{data}
\name{child_group}
\alias{child_group}
\title{Child HCC Grouping Factors - Table 9}
\format{\describe{
\item{group}{Group label for risk scores}
\item{used}{Whether or not the category is used in the RA model}
\item{plat, gold, silver, bronze, cat}{HCC grouping risk score for each
metal level}
}}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
child_group
}
\description{
A dataset containing the child model HCC grouping factors for the risk
adjustment model. Some condition categories have been grouped together, so
that any number of HCCs within that group are zeroed out and replaced by the
group.
}
\seealso{
Other factor datasets: 
\code{\link{adult_demo}},
\code{\link{adult_enroll_dur}},
\code{\link{adult_group}},
\code{\link{adult_hcc}},
\code{\link{adult_interaction}},
\code{\link{adult_rxc_hcc_inter}},
\code{\link{adult_rxc}},
\code{\link{child_demo}},
\code{\link{child_hcc}},
\code{\link{infant_demo}},
\code{\link{infant_mat_sev}}
}
\concept{factor datasets}
\keyword{datasets}
