% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefs}
\alias{coefs}
\title{Model Coefficients - Table 9}
\format{\describe{
\item{model}{Adult, Child, or Infant model, which all have different
variables and different coefficients}
}}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
coefs
}
\description{
A dataset containing the risk adjustment coefficients. Risk varies by sex,
age, enrollment duration, condition category, prescription drug category,
and adult/child/infant status. There are also some drug/diagnosis
interaction terms, and some "high risk" interaction terms.
}
\keyword{datasets}
