% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpt_hcpcs}
\alias{cpt_hcpcs}
\title{RA-eligible CPT and HCPCS - Table 2}
\format{A data frame with ~7000 rows and 5 variables:
\describe{
  \item{code}{CPT/HCPCS code}
  \item{desc}{Short description of the code}
  \item{prior}{Whether or not the code was valid in the prior benefit year}
  \item{curr}{Whether or not the code is valid in the current benefit year}
  \item{footnote}{Any footnotes included in the CMS DIY file}
}}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
cpt_hcpcs
}
\description{
A dataset containing the Current Procedural Terminology
and Health Care Procedure Coding System codes that
determine if a claim is eligible for Risk Adjustment
}
\details{
cpt_hcpcs.
}
\keyword{datasets}
