% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_cc}
\alias{icd_cc}
\title{ICD to Condition Category Crosswalk - Table 3}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8528 rows and 11 columns.}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
icd_cc
}
\description{
ICD to Condition Category Crosswalk - Table 3
}
\keyword{datasets}
