% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc.R
\docType{data}
\name{ndc_rxc}
\alias{ndc_rxc}
\title{NDC to Rx Condition Crosswalk - Table 10a}
\format{\describe{
\item{rxc}{Rx condition number without leading zeroes}
\item{desc}{Rx condition description}
\item{ndc}{NDC that maps to \code{rxc}}
}}
\source{
Most recent DIY entry at:
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
ndc_rxc
}
\description{
A dataset containing the NDC to Rx condition mappings.
}
\seealso{
Other crosswalks: 
\code{\link{hcpcs_rxc}},
\code{\link{icd_cc}}

Other RXC tables: 
\code{\link{adult_rxc_hcc_inter}},
\code{\link{adult_rxc}},
\code{\link{hcpcs_rxc}},
\code{\link{rxc_hier}}
}
\concept{RXC tables}
\concept{crosswalks}
\keyword{datasets}
