% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rxc_hier}
\alias{rxc_hier}
\title{Rx Condition Hierarchies - Table 11}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
rxc_hier
}
\description{
Rx Condition Hierarchies - Table 11
}
\keyword{datasets}
