% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\docType{data}
\name{adult_interaction}
\alias{adult_interaction}
\title{Adult Interaction Factors - Table 9}
\format{\describe{
\item{int_group}{Interaction group (M/H)}
\item{used}{Whether or not the category is used in the RA model}
\item{plat, gold, silver, bronze, cat}{HCC grouping risk score for each
metal level}
}}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
adult_interaction
}
\description{
A dataset containing the adult model interaction factors for the risk
adjustment model. Certain HCCs are considered "high risk" HCCs, and when
one of these high risk HCCs interacts with certain other HCCs, an additional
interaction term is added to the risk score.
}
\seealso{
Other factor datasets: 
\code{\link{adult_demo}},
\code{\link{adult_enroll_dur}},
\code{\link{adult_group}},
\code{\link{adult_hcc}},
\code{\link{adult_rxc_hcc_inter}},
\code{\link{adult_rxc}},
\code{\link{child_demo}},
\code{\link{child_group}},
\code{\link{child_hcc}},
\code{\link{infant_demo}},
\code{\link{infant_mat_sev}}
}
\concept{factor datasets}
\keyword{datasets}
