% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_hier}
\alias{cc_hier}
\title{Condition Category Hierarchies - Table 4}
\format{\describe{
\item{cc}{HHS condition category code}
\item{desc}{HHS condition category description}
\item{set_0}{HHS condition category that falls below \code{cc} on the hierarchy}
}}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
cc_hier
}
\description{
A dataset used to apply the risk adjustment condition category hierarchies.
The HHS risk model uses \emph{hierarchichal} condition categories. This
means that when an individual has e.g. HCC 8 (Metastatic Cancer), they do
not also get the risk score for less serious cancer categories.
}
\keyword{datasets}
