% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_coefs.R
\docType{data}
\name{child_demo}
\alias{child_demo}
\title{Child Demographic Factors - Table 9}
\format{\describe{
\item{used}{Whether or not the category is used in the RA model}
\item{plat, gold, silver, bronze, cat}{Demographic risk score for each
metal level}
\item{sex}{Male/Female demographic category}
\item{age_min, age_max}{Maximum and minimum age for each risk score}
}}
\source{
Most recent DIY entry at:
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
child_demo
}
\description{
A dataset containing the child model demographic factors for the risk
adjustment model. Demographics vary by age (groupings) and sex.
}
\seealso{
Other factors: 
\code{\link{adult_demo}},
\code{\link{adult_enroll_dur}},
\code{\link{adult_group}},
\code{\link{adult_hcc}},
\code{\link{adult_interaction}},
\code{\link{adult_rxc_hcc_inter}},
\code{\link{adult_rxc}},
\code{\link{child_group}},
\code{\link{child_hcc}},
\code{\link{infant_demo}},
\code{\link{infant_mat_sev}}
}
\concept{factors}
\keyword{datasets}
