% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefs}
\alias{coefs}
\title{Model Coefficients - Table 9}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 379 rows and 8 columns.}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
coefs
}
\description{
Model Coefficients - Table 9
}
\keyword{datasets}
