% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc.R
\docType{data}
\name{hcpcs_rxc}
\alias{hcpcs_rxc}
\title{HCPCS to Rx Condition Crosswalk - Table 10b}
\format{\describe{
\item{rxc}{Rx condition number without leading zeroes}
\item{desc}{Rx condition description}
\item{ndc}{HCPCS code that maps to \code{rxc}}
}}
\source{
Most recent DIY entry at:
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
hcpcs_rxc
}
\description{
A dataset containing the HCPCS to Rx condition mappings.
}
\seealso{
Other crosswalks: 
\code{\link{icd_cc}},
\code{\link{ndc_rxc}}

Other RXC tables: 
\code{\link{adult_rxc_hcc_inter}},
\code{\link{adult_rxc}},
\code{\link{ndc_rxc}},
\code{\link{rxc_hier}}
}
\concept{RXC tables}
\concept{crosswalks}
\keyword{datasets}
