% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcpcs_rxc}
\alias{hcpcs_rxc}
\title{HCPCS to Rx Condition Crosswalk - Table 10b}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43 rows and 3 columns.}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
hcpcs_rxc
}
\description{
HCPCS to Rx Condition Crosswalk - Table 10b
}
\keyword{datasets}
