% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc.R
\docType{data}
\name{icd_cc}
\alias{icd_cc}
\title{ICD to Condition Category Crosswalk - Table 3}
\format{\describe{
\item{icd}{ICD-10 code}
\item{cc}{Risk adjustment condition category}
\item{class}{It's not yet clear how this is used}
\item{eff_date}{Effective date of ICD code}
\item{term_date}{Term date of ICD code}
\item{sex}{Male/Female specification for ICD code (Risk adjustment
currently only supports male and female)}
\item{sex_num}{Numeric sex encoding (1/M and 2/F)}
}}
\source{
#' Most recent DIY entry at: #' \url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
icd_cc
}
\description{
A dataset containing the ICD-10 to condition category mappings.
}
\seealso{
Other crosswalks: 
\code{\link{hcpcs_rxc}},
\code{\link{ndc_rxc}}

Other HCC tables: 
\code{\link{adult_group}},
\code{\link{adult_hcc}},
\code{\link{adult_interaction}},
\code{\link{adult_rxc_hcc_inter}},
\code{\link{cc_hier}},
\code{\link{child_group}},
\code{\link{child_hcc}},
\code{\link{cpt_hcpcs}}
}
\concept{HCC tables}
\concept{crosswalks}
\keyword{datasets}
