% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infant_coefs.R
\docType{data}
\name{infant_demo}
\alias{infant_demo}
\title{Infant Demographic Factors - Table 9}
\format{\describe{
\item{used}{Whether or not the category is used in the RA model}
\item{plat, gold, silver, bronze, cat}{Demographic risk score for each
metal level}
\item{age}{Age for each risk score}
\item{sex}{Male/Female demographic category}
}}
\source{
Most recent DIY entry at:
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
infant_demo
}
\description{
A dataset containing the infant model demographic factors for the risk
adjustment model. Demographics vary by age and sex.
}
\seealso{
Other factors: 
\code{\link{adult_demo}},
\code{\link{adult_enroll_dur}},
\code{\link{adult_group}},
\code{\link{adult_hcc}},
\code{\link{adult_interaction}},
\code{\link{adult_rxc_hcc_inter}},
\code{\link{adult_rxc}},
\code{\link{child_demo}},
\code{\link{child_group}},
\code{\link{child_hcc}},
\code{\link{infant_mat_sev}}
}
\concept{factors}
\keyword{datasets}
