% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndc_rxc}
\alias{ndc_rxc}
\title{NDC to Rx Condition Crosswalk - Table 10a}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12323 rows and 3 columns.}
\source{
\url{https://www.cms.gov/cciio/Resources/Regulations-and-Guidance/index}
}
\usage{
ndc_rxc
}
\description{
NDC to Rx Condition Crosswalk - Table 10a
}
\keyword{datasets}
